perf: Refactor to run boot migration check as separate process
